// src/app/types.d.ts
export interface ImageInfo {
  id?: string;
  url?: string;
  type?: string;
  createdAt?: string;
  compressedImg?: string;
  fileName?: string;
  originalSize?: number;
}

export interface Price {
  value: number;
  currency: string;
}

export interface Coordinates {
  longitude: number;
  latitude: number;
}

export interface Photo {
  id: string;
  orderNumber: number;
  image: ImageInfo;
  type: string;
  isCover: boolean;
}
// ...
export interface Feature {
  id: string;
  name: string;
  category: string;
}

export interface GeneralInfo {
  id: string;
  slug: string;
  features: Feature[];
  make: string;
  model: string;
  price: Price;
  hullNumber: string;
  year: number;
  type: string;
  availableIn: string | null;
  address: string;
  country: string;
  city: string;
  state: string;
  coordinates: Coordinates;
  description: string;
  saleStatus: string;
  isCaptainMaintained: boolean;
  isNeverCharged: boolean;
  isLikeNew: boolean;
  isExclusive: boolean;
  isAvailableForCoBrokerage: boolean;
  isBrandConnectedToOwner: boolean;
  isModelPhotosUsed: boolean;
  video: Media;
  photos: {
    exterior: Photo[];
    interior: Photo[];
    deckLayout: Photo[];
  };
  modelSlug: string;
  brandId: string;
  vesselModelId: string;
}
// ...
interface Media {
  id: string;
  url: string;
  type: string;
  createdAt: string;
  fileName: string;
  originalSize: number;
  compressedImg: string;
}